Handle multiple configurations of the zlib library.
We need to protect $ZLIB_LIBRARIES with enclosing quotes otherwise
only the first element in the list is passed to ZLIB_LIBRARY.
This is problematic if there are multiple config libraries listed
in $ZLIB_LIBRARIES, such as when using a multi-config generator
like Visual Studio on Windows.

This is a manual cherry-pick of
60ce115a1243781e8ac7a51bd31b4ccc0db44e26.  An automatic cherry-pick
was not possible do to conflicts.